royalty: make royal chandelier suitable for throne rooms (in addition to vanilla braziers). Lower-tier nobles will also accept normal chandeliers or candelabra.
This uses the new IndustrialAge.RoomRequirement_ThingCountAnyOf class, but if another mod tries something similar (and doesn't name the class exactly as we did, with a compatible feature set), we may be incompatible with them. This patch is at least written in a way that ignores the namespace, so we can introduce the same class in other mods without restriction on load order.